requirements.txt
gymnasium==0.29.1
stable-baselines3==2.3.2
torch>=2.2.0
pygame==2.5.2
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
imageio>=2.34
imageio-ffmpeg>=0.4.9
moviepy==1.0.3
tqdm>=4.66
gradio>=4.44
Pillow>=10.0
